calculadora modificada para float

+++ src/br/com/dio/calculadora/Calculadora.java
@@ -7,7 +7,7 @@ public class Calculadora {
 	public static void main(String[] args) {
 		Scanner scan = new Scanner(System.in);
 		
-		int a, b;
+		float a, b;
 		
 		System.out.println("Digite o primeiro valor: ");
 		a = scan.nextInt();
@@ -15,9 +15,9 @@ public class Calculadora {
 		System.out.println("Digite o segundo valor: ");
 		b = scan.nextInt();
 		
-		int soma = soma(a,b);
-		int subtracao = subtracao(a,b);
-		int divisao = divisao(a,b);
+		float soma = soma(a,b);
+		float subtracao = subtracao(a,b);
+		float divisao = divisao(a,b);
 		int multiplicacao = multiplicacao(a,b);
 		
 		System.out.println("soma " + soma);
@@ -26,19 +26,19 @@ public class Calculadora {
 		System.out.println("Mult " + multiplicacao);
 	}
 
-	public static int soma(int a, int b) {
+	public static float soma(float a, float b) {
 		return a + b;
 	}
 	
-	public static int subtracao(int a, int b) {
+	public static float subtracao(float a, float b) {
 		return a - b;
 	}
 	
-	public static int divisao(int a, int b) {
+	public static float divisao(float a, float b) {
 		return a / b;
 	}
 	
-	public static int multiplicacao(int a, int b) {
+	public static float multiplicacao(float a, float b) {
 		return a * b;
 	}
